requirement.txt
flask
pdfkit
google-genai
gunicorn